Move test plugin specific functions to runtime/test package
